+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - ArcGIS Web Map element to handle optional web map & basemap Ids
+- ArcGIS Web Map element to show/hide widgets based on element config
 
 ## [8.1.0] - 2025-04-04